+++ scipy/utils/array_compatibility.py
